add .gitignore refactor models fix variable types add constructors add new model: validator_description.go

+++ broker/model/staking_params.go
@@ -6,11 +6,11 @@ type (
 	sParams struct {
 		UnbondingTime time.Duration `json:"unbonding_time"`
 		// max_validators is the maximum number of validators.
-		MaxValidators uint32 `json:"max_validators"`
+		MaxValidators uint64 `json:"max_validators"`
 		// max_entries is the max entries for either unbonding delegation or redelegation (per pair/trio).
-		MaxEntries uint32 `json:"max_entries"`
+		MaxEntries uint64 `json:"max_entries"`
 		// historical_entries is the number of historical entries to persist.
-		HistoricalEntries uint32 `json:"historical_entries"`
+		HistoricalEntries uint64 `json:"historical_entries"`
 		// bond_denom defines the bondable coin denomination.
 		BondDenom string `json:"bond_denom"`
 		// min_commission_rate is the chain-wide minimum commission rate that a validator can charge their delegators
@@ -22,7 +22,7 @@ type (
 	}
 )
 
-func NewStakingParams(height int64, maxValidators, maxEntries, historicalEntries uint32, bondDenom string,
+func NewStakingParams(height int64, maxValidators, maxEntries, historicalEntries uint64, bondDenom string,
 	minCommissionRate float64, unbondingTime time.Duration) StakingParams {
 
 	return StakingParams{
